+ change type int -> double for volume

+++ src/main/java/com/nce/backend/cars/domain/entities/Car.java
@@ -30,7 +30,7 @@ public class Car {
 
     private String engineType;
 
-    private Integer engineVolume;
+    private Double engineVolume;
 
     private String bodywork;
