+++ src/main/java/com/nce/backend/cars/infrastructure/jpa/CarJpaEntity.java
@@ -41,7 +41,7 @@ public class CarJpaEntity {
 
     private String engineType;
 
-    private Integer engineVolume;
+    private Double engineVolume;
 
     private String bodywork;
