+++ src/main/java/com/nce/backend/cars/ui/requests/AddCarAdminRequest.java
@@ -33,7 +33,7 @@ public record AddCarAdminRequest(
 
         @PositiveOrZero(message = "Engine volume cannot be less then 0")
         @NotNull
-        Integer engineVolume,
+        Double engineVolume,
 
         @NotBlank(message = "Bodywork cannot be null or blank")
         String bodywork,
