+++ src/main/java/com/nce/backend/cars/ui/requests/UpdateCarRequest.java
@@ -34,7 +34,7 @@ public record UpdateCarRequest(
 
         @PositiveOrZero(message = "Engine volume cannot be less then 0")
         @NotNull
-        Integer engineVolume,
+        Double engineVolume,
 
         @NotBlank(message = "Bodywork cannot be null or blank")
         String bodywork,
